move W_Void to the new primitive.py
