+++ spy/tests/compiler/test_operator_attr.py
@@ -1,9 +1,10 @@
 import pytest
 from spy.fqn import QN
+from spy.vm.primitive import W_Void
 from spy.vm.b import B
 from spy.vm.object import spytype, Member, Annotated
 from spy.vm.builtin import builtin_func
-from spy.vm.w import W_Type, W_Object, W_Dynamic, W_Str, W_I32, W_Void
+from spy.vm.w import W_Type, W_Object, W_Dynamic, W_Str, W_I32
 from spy.vm.opimpl import W_OpImpl, W_OpArg
 from spy.vm.registry import ModuleRegistry
 from spy.vm.vm import SPyVM
