+++ spy/tests/vm/test_vm.py
@@ -1,10 +1,11 @@
 import fixedint
 import pytest
+from spy.vm.primitive import W_Void
 from spy.vm.vm import SPyVM
 from spy.vm.b import B
 from spy.fqn import QN, FQN
 from spy.errors import SPyTypeError
-from spy.vm.object import W_Object, W_Type, spytype, W_Void, W_I32, W_Bool
+from spy.vm.object import W_Object, W_Type, spytype, W_I32, W_Bool
 from spy.vm.str import W_Str
 from spy.vm.function import W_BuiltinFunc
 from spy.vm.module import W_Module
