+++ spy/vm/function.py
@@ -3,7 +3,8 @@ from typing import TYPE_CHECKING, Any, Optional, Callable, Sequence
 from spy import ast
 from spy.ast import Color
 from spy.fqn import QN
-from spy.vm.object import W_Object, W_Type, W_Void
+from spy.vm.primitive import W_Void
+from spy.vm.object import W_Object, W_Type
 if TYPE_CHECKING:
     from spy.vm.vm import SPyVM
     from spy.vm.list import W_List
