+++ spy/vm/module.py
@@ -1,7 +1,8 @@
 from typing import TYPE_CHECKING, Optional, Iterable
 from spy.fqn import QN, FQN
+from spy.vm.primitive import W_Void
 from spy.vm.b import B
-from spy.vm.object import W_Object, spytype, W_Type, W_Dynamic, W_Void
+from spy.vm.object import W_Object, spytype, W_Type, W_Dynamic
 from spy.vm.str import W_Str
 from spy.vm.function import W_ASTFunc
 from spy.vm.builtin import builtin_func
