+++ spy/vm/modules/unsafe/ptr.py
@@ -2,9 +2,10 @@ from typing import TYPE_CHECKING, ClassVar, Optional, no_type_check
 import fixedint
 from spy.errors import SPyPanicError
 from spy.fqn import QN
+from spy.vm.primitive import W_Void
 from spy.vm.b import B
 from spy.vm.object import spytype
-from spy.vm.w import W_Object, W_I32, W_Type, W_Void, W_Str, W_Dynamic, W_Func
+from spy.vm.w import W_Object, W_I32, W_Type, W_Str, W_Dynamic, W_Func
 from spy.vm.opimpl import W_OpImpl, W_OpArg
 from spy.vm.builtin import builtin_func
 from . import UNSAFE
